+++ Dominio/Service/ServicosService.cs
@@ -2,6 +2,7 @@
 using Dominio.Modelos;
 using Dominio.Validadores;
 using FluentValidation.Results;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Dominio.Service;
 
@@ -9,11 +10,17 @@ public class ServicosService : BaseService, IServicosService
 {
     private readonly IServicosRepository _servicosRepository;
     private readonly ServicosValidator _servicosValidator;
+    private readonly IProfissionaisRepository _profissionaisRepository;
+    private readonly IAgendamentosService _agendamentosService;
 
-    public ServicosService(IServicosRepository servicosRepository, ServicosValidator servicosValidator)
+    public ServicosService(IServicosRepository servicosRepository, ServicosValidator servicosValidator, 
+        IProfissionaisRepository profissionaisRepository, 
+        IAgendamentosService agendamentosService)
     {
         _servicosRepository = servicosRepository;
         _servicosValidator = servicosValidator;
+        _profissionaisRepository = profissionaisRepository;
+        _agendamentosService = agendamentosService;
     }
 
     public async Task<Servico> Adicionar(Servico servico)
@@ -29,6 +36,11 @@ public class ServicosService : BaseService, IServicosService
             }
             throw new Exception(erros);
         }
+        if (!(await ExisteProfissionalAsync(servico.ProfissionalId)))
+        {
+            throw new Exception("Profissional não existe.");
+        }
+
         await _servicosRepository.Adicionar(servico);
         return servico;
     }
@@ -46,13 +58,22 @@ public class ServicosService : BaseService, IServicosService
             }
             throw new Exception(erros);
         }
-        var clienteBusca = await _servicosRepository.ObterPorId(servico.Id);
-        if (clienteBusca == null)
+        var servicoBusca = await _servicosRepository.ObterPorId(servico.Id);
+        if (servicoBusca == null)
         {
-            throw new Exception("Cliente não encontrada!");
+            throw new Exception("Serviço não encontrada!");
+        }
+        if (!(await ExisteProfissionalAsync(servico.ProfissionalId)))
+        {
+            throw new Exception("Profissional não existe.");
         }
         await _servicosRepository.Atualizar(servico.Id, servico);
         return true;
+    }
+
+    public async Task<bool> ExisteServicoPorProfissionalId(int id)
+    {
+        return await _servicosRepository.ExisteServicoPorProfissionalId(id);
     }
 
     public async Task<Servico> ObterPorId(int id)
@@ -74,10 +95,21 @@ public class ServicosService : BaseService, IServicosService
     {
         if (await _servicosRepository.ObterPorId(id) == null)
         {
-            throw new Exception(" não existe!");
+            throw new Exception("Serviço não existe!");
+        }
+
+        if(await _agendamentosService.ExisteServico(id))
+        {
+            throw new Exception("Serviço vinculado a um agendamento!");
         }
 
         await _servicosRepository.Remover(id);
         return true;
+    }
+
+    private async Task<bool> ExisteProfissionalAsync(int id)
+    {
+        var retorno = await _profissionaisRepository.ObterPorId(id);
+        return retorno != null;
     }
 }
